add gold dataset
rds file that has basic info for ~19k articles from several leading forestry journals. every citation has DOI, which can be used for master linkage, so we should be able to do a "best case" search test using title,author,date,journal and measure the accuracy using the DOI as key. Note that we can also make these data much bigger -- this is a 5 year cut from ~10 journals. The web of science api seems broken -- clarivate sucks -- and the microsoft academic graph tool is being phased out. But, if we think this could be useful for adding to search index as well, we can host our own model in azure and query a bunch.